improve make_rpm.sh
- download tar.gz only if not present on disk
- drop rpm_publish
- copy rpm file to top directory

+++ make_rpm.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 VERSION=$(grep "Version:" tomcat8.spec |cut -d ":" -f2 |tr -d "[:space:]")
 RELEASE=$(grep "Release:" tomcat8.spec |cut -d ":" -f2 |tr -d "[:space:]")
 ARCH=$(grep "BuildArch:" tomcat8.spec |cut -d ":" -f2 |tr -d "[:space:]")
@@ -12,7 +14,7 @@ mkdir rpmbuild/SOURCES
 mkdir rpmbuild/SPECS
 mkdir rpmbuild/SRPMS
 
-wget http://archive.apache.org/dist/tomcat/tomcat-8/v$VERSION/bin/apache-tomcat-$VERSION.tar.gz -O apache-tomcat-$VERSION.tar.gz
+[ -f apache-tomcat-$VERSION.tar.gz ] || wget http://archive.apache.org/dist/tomcat/tomcat-8/v$VERSION/bin/apache-tomcat-$VERSION.tar.gz -O apache-tomcat-$VERSION.tar.gz
 
 ln -v -s "$(pwd)/apache-tomcat-$VERSION.tar.gz" "rpmbuild/SOURCES/"
 ln -v -s "$(pwd)/tomcat8."{init,logrotate,sysconfig,bin,conf} "rpmbuild/SOURCES/"
@@ -22,5 +24,6 @@ cd rpmbuild
 
 rpmbuild --buildroot "`pwd`/BUILDROOT" ../tomcat8.spec -bb --define "_topdir `pwd`"
 
-publish-rpm $VERSION $RELEASE $ARCH tomcat8 "RPMS/$ARCH/tomcat8-$VERSION-$RELEASE.$ARCH.rpm"
+
+cp RPMS/$ARCH/tomcat8-$VERSION-$RELEASE.$ARCH.rpm ../
